+++ tests/data/depth_test.tif
